+++ src/main/java/nextstep/subway/handler/GlobalExceptionHandler.java
@@ -4,7 +4,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
-import javax.persistence.NoResultException;
 import java.util.NoSuchElementException;
 
 @RestControllerAdvice
