+++ src/test/java/nextstep/subway/line/domain/SectionsTest.java
@@ -1,8 +1,5 @@
-package nextstep.subway.sections;
+package nextstep.subway.line.domain;
 
-import nextstep.subway.line.domain.Distance;
-import nextstep.subway.line.domain.Section;
-import nextstep.subway.line.domain.Sections;
 import nextstep.subway.station.domain.Station;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
